Eindoplevering
Dit is de eindoplevering.

+++ Oplevering Iteratie 4/BasketbalAppInversion/Logic/Player.cs
@@ -16,10 +16,10 @@ namespace APPBasketbal.Models
         public string LastName { get; set; }
         public string FirstName { get; set; }
         public string ActiveTeam { get; set; }
-        public int Points { get; set; }
-        public int Assists { get; set; }
-        public int Rebounds { get; set; }
-        public int Blocks { get; set; }
+        public double Points { get; set; }
+        public double Assists { get; set; }
+        public double Rebounds { get; set; }
+        public double Blocks { get; set; }
         public string Nationality { get; set; }
     }
 }
