+++ Oplevering Iteratie 4/BasketbalAppInversion/Logic/PlayerInfo.cs
@@ -7,10 +7,10 @@ namespace Logic
     public class PlayerInfo
     {
         public int IdPlayer { get; set; }
-        public int Points { get; set; }
-        public int Rebounds { get; set; }
-        public int Assists { get; set; }
-        public int Blocks { get; set; }
+        public double Points { get; set; }
+        public double Rebounds { get; set; }
+        public double Assists { get; set; }
+        public double Blocks { get; set; }
         public string Nationality { get; set; }
     }
 }
